fix several things
PS2: to PS2 collides with swap. Switched to PS2 to ADB
Touchpad: USTP does not have to be at (\). Defined it inside the I2C device.
X510UA: IANE is newly defined inside this SSDT. Removed external declaration
_OSI-XINI: Update comment. Doesn't seem to matter if location is (\) or (_SB.PCI0) because there's only one XINI in the whole ACPI. Will probably make a custom device at root level and define XINI there so ACPI doesn't break at all (both compiling and booting)
config.plist's: Delete USTP to XSTP because with the updated Touchpad SSDT, USTP is not defined at (\)