Rename low battery SSDT, back to sleep at 5, use design capacity
There was a typo in the condition which did not work anyways with arithmetic inside it, but I tested with the corrected SSDT and the laptop does go into hibernation (at 70 percent which I set) and does not wake until I give it some charge. Use design capacity instead of last full capacity because I can't really tell when the laptop goes into hibernation with last full charge. Design capacity is a fixed value which I can set whichever value I want and when exactly I want the laptop to go into hibernation.