Hibernate at 10 percent battery level (#39)
* test sleep at assign voltage

* Fix If (One) condition and delete test at high level

* Separate sleep at low battery from main battery SSDT

* BIXT may not have data yet so execute _BIX to fill BIXT for _BST to use

* Test if it works at 70 % of last full capacity

* Rename low battery SSDT, back to sleep at 5, use design capacity

There was a typo in the condition which did not work anyways with arithmetic inside it, but I tested with the corrected SSDT and the laptop does go into hibernation (at 70 percent which I set) and does not wake until I give it some charge. Use design capacity instead of last full capacity because I can't really tell when the laptop goes into hibernation with last full charge. Design capacity is a fixed value which I can set whichever value I want and when exactly I want the laptop to go into hibernation.

* Battery Doc sleep at low battery information

* Docs: Split current battery README

* Avoid using spaces in file names

* Fix typos in ECEnabler.md

* Don't execute _BIX

* Add hibernate at low battery and set the sleep battery level at 10 percent

* typos

* Rename Modified-RECB-and-WECB-methods.md to modified-RECB-and-WECB-methods.md

* typos

* It does work without hibernation. tested on AtivBook

* Update hibernate-at-low-battery.md

* It works in Windows

* title

* title match in readme

* Windows again and about kext

* Typos and order of docs

* Upload a low battery example SSDT

* readability

* link to example

* wrong example

* config.plist rename and Operating System

* Update hibernate-at-low-battery-level.md

* readability

* Update hibernate-at-low-battery-level.md

* final touches